Add everything

+++ G53GRACW/Windmill.h
@@ -33,7 +33,7 @@ private:
 	bool toTextureB = false, toTextureR = false, toTextureW = false, toTextureF = false;
 	GLuint texidB, texidR, texidW, texidF;
 
-	float speed = 0.0074444;
+	double speed = 0.0074444;
 	int rotors = 6;
 };
 #endif
